made major changes
