+++ src/index.js
@@ -3,12 +3,26 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import { BrowserRouter as Router,Routes,Route} from "react-router-dom";
+import Navbar from './Nav';
+import Login from './login';
+import Register from './register';
+import Apply from './apply';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <App />
+    <Router>
+      <Navbar/>
+      <Routes>
+        <Route path='/' element={<App />}/>
+        <Route path='/login' element={<Login/>}/>
+        <Route path='/register' element={<Register/>}/>
+        <Route path='/apply' element={<Apply/>}/>
+      </Routes>
+    </Router>
   </React.StrictMode>
+  
 );
 
 // If you want to start measuring performance in your app, pass a function
